feat: 드래그 관련 함수 utils 함수로 재사용
