+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as formatTimeLikeTimer } from './formatTimeLikeTimer';
 export { default as formatTimeToSelect } from './formatTimeToSelect';
+export { default as handleDragEnd } from './handleDragEnd';
